src/main/webapp/app/entities/genetique/delete/genetique-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { GenetiqueService } from '../service/genetique.service';

import { GenetiqueDeleteDialogComponent } from './genetique-delete-dialog.component';

describe('Genetique Management Delete Component', () => {
  let comp: GenetiqueDeleteDialogComponent;
  let fixture: ComponentFixture<GenetiqueDeleteDialogComponent>;
  let service: GenetiqueService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [GenetiqueDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(GenetiqueDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(GenetiqueDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(GenetiqueService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
